archive/legacy_refactor_2026/requirements_minimal.txt
# Minimal requirements for Volume 2 Premiere
# Compatible with Python 3.14

# Core dependencies
pydantic>=2.0
loguru>=0.7.0
numpy>=1.24.0

# Image processing (use compatible version)
Pillow>=9.0.0

# GUI (built-in with Python)
# tkinter - included with Python
